Nest Grape::Path inside Grape::Router::Pattern
Grape::Path only ever existed to compute the origin/suffix strings that
Grape::Router::Pattern compiles into a matcher — it had a single caller
(Endpoint#to_routes) yet sat at the top level of the Grape namespace as if
it were public.

Move it to Grape::Router::Pattern::Path and add a Pattern.build factory that
takes the raw path, namespace and inheritable settings and constructs the
Path internally, so the endpoint makes one call instead of building a Path
and threading its origin/suffix into Pattern.new. Pattern.new stays
value-based (origin:/suffix:/...), so the matcher never learns about the
settings-hash shape; only Path (and the thin factory) is settings-aware.

Grape::Path is kept as a deprecated constant proxy to the new location.
