+++ lib/grape/router/pattern.rb
@@ -13,6 +13,14 @@ module Grape
       def_delegators :to_regexp, :===
       alias match? ===
 
+      # Build a Pattern from a raw path, namespace and the API's inheritable
+      # settings. {Path} owns the settings-aware assembly of +origin+/+suffix+;
+      # the Pattern itself stays value-based (see {#initialize}).
+      def self.build(path:, namespace:, settings:, anchor:, params:, version:, requirements:)
+        built_path = Path.new(path, namespace, settings)
+        new(origin: built_path.origin, suffix: built_path.suffix, anchor:, params:, version:, requirements:)
+      end
+
       def initialize(origin:, suffix:, anchor:, params:, version:, requirements:)
         @origin = origin
         @anchor = anchor
