+++ spec/grape/router/pattern_spec.rb
@@ -12,4 +12,15 @@ RSpec.describe Grape::Router::Pattern do
       expect(pattern.requirements).to eq(id: /\d+/)
     end
   end
+
+  describe '.build' do
+    subject(:pattern) do
+      described_class.build(path: '/foo', namespace: 'ns', settings: { root_prefix: '/api' }, anchor: true, params: {}, version: nil, requirements: {})
+    end
+
+    it 'assembles origin/suffix from the path, namespace and settings via Path' do
+      expect(pattern.origin).to eq('/api/ns/foo')
+      expect(pattern.path).to eq('/api/ns/foo(.:format)')
+    end
+  end
 end
